requirements.txt
requests
python-dotenv
schedule
pywin32; platform_system == "Windows"

